requirements.txt
Django>=4.0,<5
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
numpy>=1.21
requests>=2.26
groq
